Vimscript apparently has three equality operators: `==`, `==?`, and `==#`.
For string comparisons:

`==` is either case-sensitive or case-insensitive.
`==?` is always case-sensitive.
`==#` is always case-insensitive.

None of this matters for `rainbow.vim`, since I only do integer comparisons in it, but it is clear that `==?` should be preferred by default for all equality checks and so I've changed the code to reflect that new aspect of my Vimscript code style going forward, as a matter of responsibility.