The document I got the tip about case sensitive/insensitive equality checks in Vim from was apparently worded in a way that inverted the implied meanings of the two specialized equality operators.
Thus, the correct correspondences are actually these instead, for string comparisons:

`==` is either case-sensitive or case-insensitive.
`==#` is always case-sensitive.
`==?` is always case-insensitive.

As such, I've replaced the unintended uses of `==?` with `==#` now instead.

It previously hadn't showed up because it doesn't matter for this Vimscript (`rainbow.vim`) and I had taken the wording of the document I got the tip from at its word.

I should have tested it instead of trusting the implications of what the document said.

Sorry about that if it confused/misled anyone.